Moved items to their own modules - client and callback
