+++ tests/test_main.py
@@ -3,65 +3,15 @@ Unit tests for the main module
 """
 
 import unittest
-import tempfile
-import threading
 
-from resource_allocator_client.main import *
-
-
-class CallbackServerTestCase(unittest.TestCase):
-    def test_callback_server(self):
-        code_list = []
-        thread = threading.Thread(
-            target=run_callback_server,
-            kwargs={"code_list": code_list},
-            daemon=True,
-        )
-        thread.start()
-        response = req.get(url="http://localhost:8080", params={"code": "123"})
-        thread.join(timeout=10)
-        self.assertTrue(response.ok)
-        self.assertEqual(code_list, ["123"])
-
-
-class CacheTestCase(unittest.TestCase):
-    def setUp(self):
-        self.cache = Cache(server="test")
-        self.temp_dir = tempfile.TemporaryDirectory()
-
-    def tearDown(self):
-        self.temp_dir.cleanup()
-
-    def test_init(self):
-        self.assertIsNone(self.cache.token)
-        self.assertEqual(self.cache.server, "test")
-        self.assertEqual(str(self.cache.path), "test.json")
-
-    def test_write(self):
-        self.cache.token = "123"
-        self.cache._paths[0] = Path(self.temp_dir.name)
-        self.cache.write()
-
-        file_path = Path(self.temp_dir.name) / "test.json"
-        self.assertTrue(file_path.exists())
-        with open(file_path) as cur_file:
-            data = cur_file.read()
-
-        self.assertIsNotNone(data)
-
-    def test_read(self):
-        self.cache.token = "123"
-        self.cache._paths[0] = Path(self.temp_dir.name)
-        self.cache.write()
-
-        result = self.cache.read()
-        self.assertEqual(result, {"server": self.cache.server, "token": "123"})
-        self.assertEqual(self.cache.token, "123")
+from resource_allocator_client.main import (
+    Parser,
+)
 
 
 class MakeParserTestCase(unittest.TestCase):
     def setUp(self):
-        self.parser = make_parser()
+        self.parser = Parser().make_parser()
         self.base_args = ["-s", "test", "-e", "test@example.com"]
 
     def test_create(self):
@@ -74,26 +24,24 @@ class MakeParserTestCase(unittest.TestCase):
         self.assertNotIn("id", dir(result))
         self.assertTrue(isinstance(result.data, list))
         self.assertTrue(isinstance(result.data[0], str))
-        self.assertEquals(result.data[0], "name=bla")
+        self.assertEqual(result.data[0], "name=bla")
 
 
 class ParseDataArgsTestCase(unittest.TestCase):
     def test_parse_data_args(self):
-        result = parse_data_args([
+        result = Parser.parse_data_args([
             "key=value",
             " other key = value ",
-            "some key bla %'\"==invalid123",
         ])
         self.assertEqual(
             result, {
                 "key": "value",
                 "other key": "value",
-                "some key bla %'\"": "=invalid123",
             },
         )
 
     def test_parse_data_args_special(self):
-        result = parse_data_args([
+        result = Parser.parse_data_args([
             "bool_yes=yes",
             "bool_no=no",
             "bool_true=true",
@@ -118,4 +66,4 @@ class ParseDataArgsTestCase(unittest.TestCase):
 
     def test_parse_data_args_no_equals(self):
         with self.assertRaises(ValueError):
-            result = parse_data_args(["no equals"])
+            _ = Parser.parse_data_args(["no equals"])
